+++ src/Console/InstallCommand.php
@@ -113,8 +113,8 @@ class InstallCommand extends Command
         ];
 
         $packageDevDependencies = [
-            'vite' => '^4.5.0',
-            'laravel-vite-plugin' => '^0.7.2',
+            'vite' => '^5.4.0',
+            'laravel-vite-plugin' => '^1.0.0',
         ];
 
         if (File::exists($packageJsonPath)) {